+++ src/ui/app.py
@@ -45,7 +45,7 @@ from src.traders.stock.ibkr_trader import IBKRTrader
 
 
 # Application version
-TRADLEWARE_VERSION = "v3.4.0b"
+TRADLEWARE_VERSION = "v3.4.1b"
 
 # You might need to adjust this import based on where your logger.py is relative to app.py
 # If your logger is within src/misc, you might access it like this: